Adding Agilant Workflow
This is a script developed by Rich Perez. 

It is being opened to developers on the Agilant Team.

Currently, there are two developers debugging the script:
Rich, Reion Quamina

Others from the team worth mentioning have contributed to the logic included within the script.

The changes committed involve filtering out duplicate entries collected in the combined data pandas dataframe. 

Currently, we aim to include the record ID of previous entries, include the new designator convention, and suggest corrections for the found discrepancies.

The script is going to be refactored when functionality and debugging are complete.