upload file is based on dynamic_duplicates.py
This version of the script has moved forward in factoring logic into their own methods although not to completion.

It also begins to address how to properly count floor plans and sequences. Once figured out, the logic can be used to count duplicates properly.

It is currently astronomically slow -- the slowest it's ever been. Will have to work on this without a doubt.